update
update

+++ controllers/videocontroller.js
@@ -1,5 +1,6 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import Comment from "../models/Comment";
 
 // Home
 
@@ -32,8 +33,9 @@ export const search = async (req, res) => {
 
 // Upload
 
-export const getUpload = (req, res) =>
+export const getUpload = (req, res) => {
     res.render("upload", { pageTitle: "Upload" });
+}
 
 export const postUpload = async (req, res) => {
     const {
@@ -43,8 +45,11 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -55,7 +60,9 @@ export const videoDetail = async (req, res) => {
         params: { id }
     } = req;
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id)
+            .populate("creator")
+            .populate("comments");
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -70,7 +77,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -96,9 +107,79 @@ export const deleteVideo = async (req, res) => {
         params: { id }
     } = req;
     try {
-        await Video.findOneAndRemove({ _id: id });
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch (error) {
         console.log(error);
     }
     res.redirect(routes.home);
+};
+
+// Register Video View
+
+export const postRegisterView = async (req, res) => {
+    const {
+        params: { id }
+    } = req;
+    try {
+        const video = await Video.findById(id);
+        video.views += 1;
+        video.save();
+        res.status(200);
+    } catch (error) {
+        res.status(400);
+    } finally {
+        res.end();
+    }
+};
+
+// Add Comment
+
+export const postAddComment = async (req, res) => {
+    const {
+        params: { id },
+        body: { comment },
+        user
+    } = req;
+    console.log("postAddComment: " + id);
+    try {
+        const video = await Video.findById(id);
+        const newComment = await Comment.create({
+            text: comment,
+            creator: user.id
+        });
+        video.comments.push(newComment.id);
+        video.save();
+    } catch (error) {
+        res.status(400);
+    } finally {
+        res.end();
+    }
+};
+
+//Delete Comment
+
+export const postDeleteComment = async (req, res) => {
+
+    const {
+        params: { id }
+    } = req;
+
+    try {
+        const video = await Video.findById(id);
+        const comments = await Comment.findById(id);
+        // fs.unlinkSync(comments)
+        // await Comment.findByIdAndRemove({ _id: id })
+        console.log("postDeleteComment: " + video.comments[0])
+        video.comment[0].removeChild(comments);
+        video.save();
+    } catch (error) {
+        res.status(400);
+    } finally {
+        res.end();
+    }
 };
